+++ Sources/KinescopeSDK/Events/InnerEventsHandler.swift
@@ -7,18 +7,20 @@
 
 import Foundation
 
-protocol InnerEventsHandler {
-    func playback(sec: TimeInterval)
-    func play()
-    func pause()
-    func end()
-    func replay()
-    func buffer(sec: TimeInterval)
-    func seek()
-    func rate(_ rate: Float)
-    func view()
-    func enterfullscreen()
-    func exitfullscreen()
-    func qualitychanged(_ quality: String)
-    func autoqualitychanged(_ quality: String)
+public protocol InnerEventsHandler {
+    func sendOnce(event: InnerProtoEvent, value: Float)
+    func send(event: InnerProtoEvent, value: Float)
+    func reset()
+}
+
+// MARK: - Defaults
+
+extension InnerEventsHandler {
+    func send(event: InnerProtoEvent) {
+        send(event: event, value: 0)
+    }
+
+    func sendOnce(event: InnerProtoEvent) {
+        sendOnce(event: event, value: 0)
+    }
 }