+++ Sources/KinescopeSDK/Manager/Manager+KinescopeConfigurable.swift
@@ -22,8 +22,8 @@ extension Manager: KinescopeConfigurable {
                                                assetService: AssetNetworkService())
         self.inspector = Inspector(videosService: VideosNetworkService(transport: Transport(),
                                                                        config: config))
+        self.analyticFactory = AnalyticHandler(service: AnalyticsNetworkService(transport: Transport(), config: config))
 
-        self.eventsCenter = LocalEventsCenter()
     }
 
     func set(logger: KinescopeLogging, levels: [KinescopeLoggingLevel]) {