+++ Sources/KinescopeSDK/Player/Dependencies/KinescopePlayerDependencies.swift
@@ -11,5 +11,4 @@ protocol KinescopePlayerDependencies: PlayingStrategyProvider {
     var inspector: KinescopeInspectable { get }
     var assetDownloader: KinescopeAssetDownloadable { get }
     var attachmentDownloader: KinescopeAttachmentDownloadable { get }
-    var eventsCenter: KinescopeEventsCenter { get }
 }