+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Removed
 - APIkey requirement
 - API calls including `videoList`
+- old example with many videos and scrolling
 
 ### Added
 - implementation of FairPlayStreaming (DRM)
@@ -10,3 +11,6 @@
 - automatic retry mechanism for video loading
 - error state with refresh button
 - live stream announce view
+- collecting of analytic events with playback data for Kinescope dashboard
+- textField in example project to allow user to input videoId
+